Move prompt config to mysql section
For newer versions of MariaDB the prompt config needs to be in the
mysql section of my.cnf. Having it in the client section can lead
to mysql commands failing on some versions of MariaDB (ie. on Debian
9 Stretch). More details on the `prompt` configuration can be found
in the MySQL docs: https://dev.mysql.com/doc/refman/5.7/en/mysql-commands.html